requirements.txt
paramiko>=3.4.0
pyserial>=3.5